Services/api/requirements.txt
fastapi==0.128.5
uvicorn==0.40.0
httpx==0.28.1
minio==7.2.20
python-multipart==0.0.22
prometheus-fastapi-instrumentator==7.1.0
slowapi==0.1.9
redis==7.3.0
pydantic-settings==2.13.1
sqlalchemy==2.0.48
alembic==1.16.5
psycopg2-binary==2.9.11
argon2-cffi==25.1.0
pyjwt==2.12.1
email-validator==2.2.0
itsdangerous==2.2.0